Delete PodDisruptionBudgets when destroying cluster
Without this change the termination lifecycle hook Lambda hangs until it
timesout as it tries to drain nodes that fail to evict pods because it
would cause a violation to pod disruption budgets.

Whilst this is sensible most of the time (for example when rolling
nodes) it's not required when destroying a cluster entirely and actually
prevents them being destroyed without intervention.